refactor pdf generation & finish api
