+++ backend/routes/api/examine/expenseReport.ts
@@ -1,28 +1,71 @@
+import { getter, setter } from '../../../helper.js'
 import express, { Request, Response } from 'express'
 const router = express.Router()
+import DocumentFile from '../../../models/documentFile.js'
+import ExpenseReport, { ExpenseReportDoc } from '../../../models/expenseReport.js'
+import i18n from '../../../i18n.js'
 import multer from 'multer'
 const fileHandler = multer({ limits: { fileSize: 16000000 } })
-import i18n from '../../i18n.js'
-import { getter, setter, deleter } from '../../helper.js'
-import ExpenseReport, { ExpenseReportDoc } from '../../models/expenseReport.js'
-import DocumentFile from '../../models/documentFile.js'
-import { sendExpenseReportNotificationMail } from '../../mail/mail.js'
-import { ExpenseReport as IExpenseReport } from '../../../common/types.js'
-import { generateExpenseReportReport } from '../../pdf/expenseReport.js'
+import { sendExpenseReportNotificationMail } from '../../../mail/mail.js'
+import { ExpenseReport as IExpenseReport } from '../../../../common/types.js'
+import { writeToDisk } from '../../../pdf/helper.js'
+import { generateExpenseReportReport } from '../../../pdf/expenseReport.js'
 
 router.get('/', async (req, res) => {
-  const sortFn = (a: IExpenseReport, b: IExpenseReport) => (a.createdAt as Date).valueOf() - (b.createdAt as Date).valueOf()
+  const sortFn = (a: IExpenseReport, b: IExpenseReport) => (a.updatedAt as Date).valueOf() - (b.updatedAt as Date).valueOf()
   const select: Partial<{ [key in keyof IExpenseReport]: number }> = { history: 0, historic: 0 }
   if (!req.query.addExpenses) {
     select.expenses = 0
   }
   delete req.query.addExpenses
-  return getter(ExpenseReport, 'expense report', 20, { expensePayer: req.user!._id, historic: false }, select, sortFn)(req, res)
+  return getter(ExpenseReport, 'expense report', 20, { state: 'underExamination', historic: false }, select, sortFn)(req, res)
 })
 
-router.delete('/', deleter(ExpenseReport, 'expensePayer'))
+router.get('/refunded', async (req, res) => {
+  const sortFn = (a: IExpenseReport, b: IExpenseReport) => (b.updatedAt as Date).valueOf() - (a.updatedAt as Date).valueOf() // sort backwards
+  const select: Partial<{ [key in keyof IExpenseReport]: number }> = { history: 0, historic: 0 }
+  if (!req.query.addExpenses) {
+    select.expenses = 0
+  }
+  delete req.query.addExpenses
+  return getter(ExpenseReport, 'expense report', 20, { state: 'refunded', historic: false }, select, sortFn)(req, res)
+})
+
+router.post('/refunded', async (req, res) => {
+  req.body = {
+    state: 'refunded',
+    editor: req.user!._id,
+    comment: req.body.comment,
+    _id: req.body._id
+  }
+  const check = async (oldObject: ExpenseReportDoc) => {
+    if (oldObject.state === 'underExamination') {
+      await oldObject.saveToHistory()
+      await oldObject.save()
+      return true
+    } else {
+      return false
+    }
+  }
+  const cb = async (expenseReport: IExpenseReport) => {
+    sendExpenseReportNotificationMail(expenseReport)
+    if (process.env.BACKEND_SAVE_REPORTS_ON_DISK.toLowerCase() === 'true') {
+      await writeToDisk(
+        '/reports/expenseReport' + expenseReport.expensePayer.name + ' - ' + expenseReport.name + '.pdf',
+        await generateExpenseReportReport(expenseReport)
+      )
+    }
+  }
+  return setter(ExpenseReport, '', false, check, cb)(req, res)
+})
 
 router.post('/expense', fileHandler.any(), async (req: Request, res: Response) => {
+  const expenseReport = await ExpenseReport.findOne({ _id: req.body.expenseReportId })
+  delete req.body.expenseReportId
+  if (!expenseReport || expenseReport.historic || expenseReport.state !== 'underExamination') {
+    return res.sendStatus(403)
+  }
+
   if (req.body.cost && req.body.cost.receipts && req.files) {
     for (var i = 0; i < req.body.cost.receipts.length; i++) {
       var buffer = null
@@ -33,21 +76,12 @@ router.post('/expense', fileHandler.any(), async (req: Request, res: Response) =
         }
       }
       if (buffer) {
-        req.body.cost.receipts[i].owner = req.user!._id
+        req.body.cost.receipts[i].owner = expenseReport.expensePayer._id
         req.body.cost.receipts[i].data = buffer
       }
     }
   }
-  const expenseReport = await ExpenseReport.findOne({ _id: req.body.expenseReportId })
-  delete req.body.expenseReportId
-  if (
-    !expenseReport ||
-    expenseReport.historic ||
-    expenseReport.state !== 'inWork' ||
-    !expenseReport.expensePayer._id.equals(req.user!._id)
-  ) {
-    return res.sendStatus(403)
-  }
+
   if (req.body._id && req.body._id !== '') {
     var found = false
     outer_loop: for (const expense of expenseReport.expenses) {
@@ -91,11 +125,10 @@ router.post('/expense', fileHandler.any(), async (req: Request, res: Response) =
     expenseReport.expenses.push(req.body)
   }
   expenseReport.expenses.sort((a, b) => new Date(a.cost.date).valueOf() - new Date(b.cost.date).valueOf())
-
   expenseReport.markModified('expenses')
   try {
-    const result = await expenseReport.save()
-    res.send({ message: i18n.t('alerts.successSaving'), result: result })
+    const result1 = await expenseReport.save()
+    res.send({ message: i18n.t('alerts.successSaving'), result: result1 })
   } catch (error) {
     res.status(400).send({ message: i18n.t('alerts.errorSaving'), error: error })
   }
@@ -104,12 +137,7 @@ router.post('/expense', fileHandler.any(), async (req: Request, res: Response) =
 router.delete('/expense', async (req: Request, res: Response) => {
   const expenseReport = await ExpenseReport.findOne({ _id: req.query.expenseReportId })
   delete req.query.expenseReportId
-  if (
-    !expenseReport ||
-    expenseReport.historic ||
-    expenseReport.state !== 'inWork' ||
-    !expenseReport.expensePayer._id.equals(req.user!._id)
-  ) {
+  if (!expenseReport || expenseReport.historic || expenseReport.state !== 'underExamination') {
     return res.sendStatus(403)
   }
   if (req.query.id && req.query.id !== '') {
@@ -141,61 +169,11 @@ router.delete('/expense', async (req: Request, res: Response) => {
   }
 })
 
-router.post('/inWork', async (req, res) => {
-  req.body = {
-    state: 'inWork',
-    expensePayer: req.user!._id,
-    editor: req.user!._id,
-    _id: req.body._id,
-    name: req.body.name
-  }
-
-  if (!req.body.name) {
-    try {
-      var date = new Date()
-      req.body.name =
-        i18n.t('labels.expenses', { lng: req.user!.settings.language }) +
-        ' ' +
-        i18n.t('monthsShort.' + date.getUTCMonth(), { lng: req.user!.settings.language }) +
-        ' ' +
-        date.getUTCFullYear()
-    } catch (error) {
-      return res.status(400).send(error)
-    }
-  }
-  return setter(ExpenseReport, 'expensePayer', true)(req, res)
-})
-
-router.post('/underExamination', async (req, res) => {
-  req.body = {
-    state: 'underExamination',
-    editor: req.user!._id,
-    comment: req.body.comment,
-    _id: req.body._id
-  }
-
-  const check = async (oldObject: ExpenseReportDoc) => {
-    if (oldObject.state === 'inWork') {
-      await oldObject.saveToHistory()
-      await oldObject.save()
-      return true
-    } else {
-      return false
-    }
-  }
-  return setter(ExpenseReport, 'expensePayer', false, check, sendExpenseReportNotificationMail)(req, res)
-})
-
 router.get('/report', async (req, res) => {
-  const expenseReport = await ExpenseReport.findOne({
-    _id: req.query.id,
-    expensePayer: req.user!._id,
-    historic: false,
-    state: 'refunded'
-  }).lean()
+  const expenseReport = await ExpenseReport.findOne({ _id: req.query.id, historic: false, state: 'refunded' }).lean()
   if (expenseReport) {
     const report = await generateExpenseReportReport(expenseReport)
-    res.setHeader('Content-disposition', 'attachment; filename=' + expenseReport.name + '.pdf')
+    res.setHeader('Content-disposition', 'attachment; filename=' + expenseReport.expensePayer.name + ' - ' + expenseReport.name + '.pdf')
     res.setHeader('Content-Type', 'application/pdf')
     res.setHeader('Content-Length', report.length)
     return res.send(Buffer.from(report))
